# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LedgerGuard AI Treasury OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  address: string;
  network: string;
  /** @nullable */
  label?: string | null;
  createdAt: string;
}

export interface WalletInput {
  address: string;
  network: string;
  label?: string;
}

export interface WalletBalance {
  address: string;
  balanceLamports: number;
  balanceSol: number;
  network: string;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  signed: 'signed',
  broadcast: 'broadcast',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  /** @nullable */
  signature?: string | null;
  amount: number;
  recipient: string;
  token: string;
  /** @nullable */
  memo?: string | null;
  status: TransactionStatus;
  /** @nullable */
  riskScore: number | null;
  /** @nullable */
  riskLevel?: string | null;
  network: string;
  /** @nullable */
  fromWalletAddress?: string | null;
  aiProposed?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface TransactionInput {
  amount: number;
  recipient: string;
  token: string;
  memo?: string;
  network: string;
  fromWalletAddress?: string;
  aiProposed?: boolean;
}

export type TransactionUpdateStatus = typeof TransactionUpdateStatus[keyof typeof TransactionUpdateStatus];


export const TransactionUpdateStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  signed: 'signed',
  broadcast: 'broadcast',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface TransactionUpdate {
  status?: TransactionUpdateStatus;
  signature?: string;
  riskScore?: number;
  riskLevel?: string;
}

export interface SimulationResult {
  success: boolean;
  estimatedFee: number;
  logs: string[];
  /** @nullable */
  error?: string | null;
}

export interface BroadcastInput {
  signedTransaction: string;
}

export interface BroadcastResult {
  signature: string;
  success: boolean;
  /** @nullable */
  error?: string | null;
}

export type RiskAssessmentLevel = typeof RiskAssessmentLevel[keyof typeof RiskAssessmentLevel];


export const RiskAssessmentLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL',
} as const;

export interface RiskAssessment {
  riskScore: number;
  level: RiskAssessmentLevel;
  reasons: string[];
}

export interface Policy {
  id: number;
  type: string;
  name?: string;
  value: string;
  enabled: boolean;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface PolicyInput {
  type: string;
  name: string;
  value: string;
  enabled?: boolean;
  description?: string;
}

export interface PolicyUpdate {
  type?: string;
  name?: string;
  value?: string;
  enabled?: boolean;
  description?: string;
}

export interface PolicyValidationInput {
  amount: number;
  recipient: string;
  token: string;
  timestamp?: string;
}

export interface PolicyValidationResult {
  passed: boolean;
  violations: string[];
  warnings?: string[];
}

export interface AiConversation {
  id: number;
  prompt: string;
  response: string;
  /** @nullable */
  actionProposal?: string | null;
  createdAt: string;
}

export interface AiMessageInput {
  prompt: string;
  walletAddress?: string;
  treasuryBalance?: number;
}

export type AiMessageResponseActionProposal = {
  action?: string;
  amount?: number;
  token?: string;
  recipient?: string;
  reason?: string;
} | null;

export interface AiMessageResponse {
  id: number;
  response: string;
  actionProposal?: AiMessageResponseActionProposal;
}

export interface AuditLog {
  id: number;
  event: string;
  /** @nullable */
  metadata?: string | null;
  /** @nullable */
  walletAddress?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  treasuryBalance: number;
  portfolioValue: number;
  monthlyOutflow: number;
  monthlyInflow: number;
  riskScore: number;
  healthScore: number;
  pendingCount: number;
  connectedWallets: number;
  network: string;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  status?: string | null;
  timestamp: string;
}

export interface SpendingDataPoint {
  label: string;
  amount: number;
  count?: number;
}

export interface SpendingAnalytics {
  period: string;
  data: SpendingDataPoint[];
  totalSpent: number;
  avgTransactionSize: number;
  transactionCount: number;
}

export interface TreasuryHealth {
  healthScore: number;
  runway: number;
  burnRate: number;
  runwayMonths: number;
  status?: string;
}

export interface RecipientSummary {
  address: string;
  /** @nullable */
  label?: string | null;
  totalAmount: number;
  transactionCount: number;
}

export type GetNonce200 = {
  nonce?: string;
};

export type SignInBody = {
  message: string;
  signature: string;
  publicKey: string;
};

export type GetMe200 = {
  address?: string;
};

export type ListTransactionsParams = {
status?: string;
limit?: number;
offset?: number;
};

export type ListAiConversationsParams = {
limit?: number;
};

export type ListAuditLogsParams = {
limit?: number;
offset?: number;
event?: string;
};

export type GetRecentActivityParams = {
limit?: number;
};

export type GetSpendingAnalyticsParams = {
period?: GetSpendingAnalyticsPeriod;
};

export type GetSpendingAnalyticsPeriod = typeof GetSpendingAnalyticsPeriod[keyof typeof GetSpendingAnalyticsPeriod];


export const GetSpendingAnalyticsPeriod = {
  weekly: 'weekly',
  monthly: 'monthly',
  quarterly: 'quarterly',
} as const;

export type GetTopRecipientsParams = {
limit?: number;
};