# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LedgerGuard AI Treasury OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signin:
    post:
      operationId: signIn
      tags: [auth]
      summary: Sign in with Solana (SIWS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, signature, publicKey]
              properties:
                message:
                  type: string
                signature:
                  type: string
                publicKey:
                  type: string
      responses:
        "200":
          description: Signed in
        "401":
          description: Invalid signature
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
        "401":
          description: Not authenticated
...
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: List treasury policies
      responses:
        "200":
          description: Policy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Create a policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyInput"
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /policies/{id}:
    patch:
      operationId: updatePolicy
      tags: [policies]
      summary: Update a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyUpdate"
      responses:
        "200":
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    delete:
      operationId: deletePolicy
      tags: [policies]
      summary: Delete a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /policies/validate:
    post:
      operationId: validateAgainstPolicies
      tags: [policies]
      summary: Validate a proposed transaction against all policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyValidationInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyValidationResult"

  # ── AI Treasury ───────────────────────────────────────────────────────────────
  /ai/conversations:
    get:
      operationId: listAiConversations
      tags: [ai]
      summary: List AI conversation history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Conversation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiConversation"
  /ai/chat:
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send a message to the AI treasury assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "200":
          description: AI response with optional action proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessageResponse"

  # ── Audit Logs ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Policy:
      type: object
      required: [id, type, value, enabled, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        name:
          type: string
        value:
          type: string
        enabled:
          type: boolean
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    PolicyInput:
      type: object
      required: [type, name, value]
      properties:
        type:
          type: string
        name:
          type: string
        value:
          type: string
        enabled:
          type: boolean
        description:
          type: string
    PolicyUpdate:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        value:
          type: string
        enabled:
          type: boolean
        description:
          type: string
    PolicyValidationInput:
      type: object
      required: [amount, recipient, token]
      properties:
        amount:
          type: number
        recipient:
          type: string
        token:
          type: string
        timestamp:
          type: string
    PolicyValidationResult:
      type: object
      required: [passed, violations]
      properties:
        passed:
          type: boolean
        violations:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string

    # ── AI ─────────────────────────────────────────────────────────────────────
    AiConversation:
      type: object
      required: [id, prompt, response, createdAt]
      properties:
        id:
          type: integer
        prompt:
          type: string
        response:
          type: string
        actionProposal:
          type: ["string", "null"]
        createdAt:
          type: string
    AiMessageInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
        walletAddress:
          type: string
        treasuryBalance:
          type: number
    AiMessageResponse:
      type: object
      required: [id, response]
      properties:
        id:
          type: integer
        response:
          type: string
        actionProposal:
          type: object
          nullable: true
          properties:
            action:
              type: string
            amount:
              type: number
            token:
              type: string
            recipient:
              type: string
            reason:
              type: string

    # ── Audit Logs ─────────────────────────────────────────────────────────────
...